Cmdlets to fetch and upload VHDs.
The Get-AzureDtlVhd cmdlet does the following:
 - Gets a specific vhd from a lab, if the -VhdName parameter is specified.
  - Gets a specific vhd from a lab, if the -VhdUri parameter is specified.
  - Gets all vhds from a lab, if the -Lab parameter is specified.

The Add-AzureDtlVhd cmdlet uploads a vhd into a lab. Please note that the vhd file must meet the following specific requirements (dictated by Azure):
 - Must be a Gen1 vhd file (and NOT a Gen2 vhdx file).
 - Fixed sized vhd (and NOT dynamically expanding vhd).
 - Size must be less than 1023 GB.
 - The vhd must be uploaded as a page blob (and NOT as a block blob).

Note: This is an initial commit. The cmdlets are a work in progress. Support for file URIs, SAS tokens etc coming soon.