Update ConfigureWinRM.ps1
This script is designed to require a value for the `hostname` input parameter. However, the parameter was not marked as mandatory. This pull request adds the [Parameter()] attribute, and designates the `-HostName` parameter as mandatory. It also fixes the casing of the parameter name.

Cheers,
Trevor Sullivan
Microsoft MVP: PowerShell