Add lint and reformat target.
This is part of the ground work to have a consistent style to the template
JSON documents.  Two targets are defined.

 * Lint - verify all JSON templates follow the approved style.
 * Reformat - modify all JSON templates to follow the approved style.

The lint target will be used to gate the build in a future commit.  If a
PR does not meet the approved style it will fail the build

The reformat target is for developers. Developers should execute this
target before sending a PR to match the approved style.