Updated Storage Account
This fixes a problem where Storage Accounts were not treated in an idempotent fashion. Updating to 2016-01-01 for the API version, just for the Storage Account, will resolve this problem. However, the schema for the Storage Account changed, so we must add the `sku` and `kind` properties at the root of the Storage Account definition.