Update 101-vsts-cloudtest-rig to fix VM deployment error
Added in a dependsOn declaration for the storage account in the virtual machine.

This fixes a deployment error that happens when the storage account isn't immediately created (e.g. taking up to 20 minutes to create to the storage account) and forces the VM to be created after the storage account has successfully been created.

Currently if the storage account isn't created immediately the deployment fails because the virtual machine requires it.