Add Spinnaker to Kubernetes template
This spins up a linux vm and a kubernetes cluster.  It installs Spinnaker on the vm and configures it to target Kubernetes.

The most complex part is getting the kube config file (used to authenticate) from the kubernetes cluster to the spinnaker vm. I use the service principal and azure cli to give a temp user access to the kubernetes master vms, then scp the file to the spinnaker vm.