Logon method RD Gateway causes Single Sign On to break
The Quick Start Template "RDS farm deployment using existing active
directory" configures the RD Gateway with the LogonMethod set to
AllowUserToSelectDuringConnection. This causes RD Web Single Sign On to
break.

Suggestion is to update the DSC Extension that is used to configure RD
Gateway and set the LogonMethod to Password