Logon method RD Gateway causes Single Sign On to break
The Quick Start Template "rds-deployment" configures the RD Gateway with
the LogonMethod set to  AllowUserToSelectDuringConnection. This causes
RD Web Single Sign On to break.

Suggestion is to update the DSC Extension that is used to configure RD
Gateway and set the LogonMethod to Password.

Issue: https://github.com/mmarch/rds-templates/issues/30