Updated Text To Better Reflect Template Use
Hi folks,  no company is going to just hand out domain administrator privileges to just anyone. I have tried to remove this requirement but I am having to dig quite deep into the weeds to get it working. 

This template is not suited for any enterprise class customer, and probably would not work for any but the smallest of IT departments. As such I've added verbiage to reflect that

Thanks!