removed unnecessary "/"
Slashes are added appropriately in the concatenation of the URI in the installScriptActions block. Leaving a trailing "/" at the end of the _artifactsLocationSasToken parameter actually results in two consecutive slashes in the URI, but it does not create an error because GitHub handles the mistake elegantly. However, this creates a somewhat misleading example.