vmss-master-slave-cluster
This template allows you to deploy multiple VM Scale Sets of Linux VMs with Master/Slave architecture. Master nodes with the lower size of VMs and data node with higher size ( Based on Elastic Search as an example, it will adapt all the master/ slave architecture ). Open to specify the numbers of master node and data node to adapt different architecture.