Default to not use prerelease releases of the agent
Currently the download of the agent does not exclude prerelease releases.
This provides the option to download prerelease or not and to default to not. This is a breaking change and so the default could be made to $true to have same behaviour. My expectation is most should not use the prerelease agent