Update location ID to Data Center location mapping
Due to historical reasons, the location ID string that is used does not always correspond to the data center from which the test is performed. I've updated the table to correctly identify the data center that is displayed in portal - which is the actual webtest location - when you create a test using ARM templates.