Added Anbox Cloud quick-start template
This commit introduces a quick-start template that facilitates
deployment of an Anbox Cloud instance. The template allows users to
launch the VM from an Ubuntu Pro image or attach an existing Ubuntu Pro
token to a VM launched from a non-Pro image. The template also supports
both the x86 and ARM processor architectures.