Added advanced configuration parameters to Anbox Cloud template
This commit adds three parameters to the Anbox Cloud quick-start
template: addDedicatedDataDiskForLXD, exposeAnboxContainerServices, and
exposeAnboxManagementService. The first allows users to specify whether
a dedicated data disk for the LXD storage pool should be attached to the
virtual machine, while the second and third expose additional ports to
the public internet. These parameters are accompanied by related content
in the README.