Anbox: Support CoreVM offers
The plan field only needs to be populated when using an AzureVM
image. Populating the field when using a CoreVM will result in an
error. This extra logic ensures the plan will only be populated when
using AzureVM and no Pro token is provided.

This change has been tested with
ubuntu-24_04-lts:ubuntu-pro-arm64 (CoreVM) and
0001-com-ubuntu-pro-jammy:pro-22_04-lts-arm64 (AzureVM).

Co-authored-by: Daniel Draper <daniel.draper@canonical.com>